requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
Django>=4.2.0
django-environ==0.12.0
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib>=3.8.0
narwhals==1.34.0
numpy>=1.24.0
packaging==24.2
pandas>=2.1.0
pillow==11.1.0
plotly>=5.18.0
psycopg2-binary>=2.9.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
seaborn>=0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
python-dotenv>=1.0.0
djangorestframework>=3.14.0
django-filter>=23.5
django-redis>=5.4.0
redis>=5.0.1
gunicorn>=21.2.0
whitenoise>=6.6.0
